Item client re-added

+++ item-service/src/main/java/com/genspark/ItemService/Service/ItemService.java
@@ -61,7 +61,7 @@ public class ItemService {
         return repo.findByScore(score);
     }
 
-    public List<Items> getByPrice(int price){
+    public List<Items> getByPrice(double price){
         return repo.findByPrice(price);
     }
     public List<Items> getByDescription(String description){
